Initial version

+++ src/MyApplication.java
@@ -1,4 +1,4 @@
-package main;
+
 import java.util.Objects;
 import java.util.Scanner;
 
@@ -9,6 +9,8 @@ import controller.UserController;
 import entities.Movie;
 
 import static java.lang.System.exit;
+
+
 public class MyApplication {
     private final UserController controller;
     private final AdminController adminController;
@@ -22,33 +24,33 @@ public class MyApplication {
     }
 
     public void start(){
-            System.out.println("""
-        Welcome to the Movie night
-        select option:
+        System.out.println("""
+        Welcome to the Movie night!
+        Please, select option:
         1. Sign up
         2. Sign in
         3. All Users
         0. Exit
         Enter option (1-3):
         """);
-            try {
+        try {
 
-                int option = scanner.nextInt();
-                switch (option){
-                    case 1 -> createUserMenu();
-                    case 2 -> SignIn();
-                    case 3 -> getAllUsersMenu();
-                    default -> exit(0);
-                }
-            } catch (Exception e) {
-                System.out.println((e.getMessage()));
-                scanner.nextLine();
+            int option = scanner.nextInt();
+            switch (option){
+                case 1 -> createUserMenu();
+                case 2 -> SignIn();
+                case 3 -> getAllUsersMenu();
+                default -> exit(0);
             }
+        } catch (Exception e) {
+            System.out.println((e.getMessage()));
+            scanner.nextLine();
         }
+    }
     public void SignIn(){
-        System.out.println("Please input your username");
+        System.out.println("Please, input your username");
         String name = scanner.next();
-        System.out.println("Please input your password");
+        System.out.println("Please, input your password");
         String pass = scanner.next();
         boolean exist = controller.SignIn(name,pass);
         if(exist){
@@ -60,7 +62,7 @@ public class MyApplication {
             else forClient();
         }
         else{
-            System.out.println("There is no such account, try again");
+            System.out.println("There is no such account, try again!");
             SignIn();
         }
 
@@ -68,7 +70,7 @@ public class MyApplication {
     public void getAllUsersMenu() throws Exception {
         String response = controller.getAllUsers();
         if(response == null){
-            System.out.println("There is no accounts, create new ");
+            System.out.println("There is no accounts, create new! ");
             createUserMenu();
         }
         else {
@@ -79,20 +81,20 @@ public class MyApplication {
 
     public void createUserMenu() throws Exception {
         String response = null;
-        System.out.println("Please enter your name");
+        System.out.println("Please, enter your name");
         String name = scanner.next();
-        System.out.println("Please enter your password");
+        System.out.println("Please, enter your password");
         String password = scanner.next();
         isRightPassword(password);
-        System.out.println("Please enter your role admin/client");
+        System.out.println("Please, enter your role admin/client");
         String choice = scanner.next();
         if(choice.equalsIgnoreCase("admin")){
             response = controller.createUsers(name, password, choice);
             forAdmin();
         } else if (choice.equalsIgnoreCase("client")) {
-            System.out.println("Please input your balance");
+            System.out.println("Please, input your balance");
             double balance = scanner.nextDouble();
-            System.out.println("Please input your age");
+            System.out.println("Please, input your age");
             int age = scanner.nextInt();
             response = controller.createUsers(name, password, choice, balance,age);
             forClient();
@@ -104,17 +106,18 @@ public class MyApplication {
     }
     public void forAdmin()  {
         System.out.println("""
+                ***********************************
                 What do you want to do?
                 1. Create movie
                 2. Update movie
                 3. Delete movie
-                4. log out from my account
+                4. Log out from my account
                 5. All movie
                 6. All users
-                7. Ban users
+                7. Delete the user's account
                 0. Exit
                 Enter option(1-7)""");
-            try {
+        try {
             int choice = scanner.nextInt();
             switch (choice) {
                 case 1 -> createMovie();
@@ -128,43 +131,52 @@ public class MyApplication {
             }
 
         }catch (Exception e) {
-                System.out.println((e.getMessage()));
-                scanner.nextLine();
-            }
+            System.out.println((e.getMessage()));
+            scanner.nextLine();
+        }
     }
     public void createMovie(){
         String movie= null;
-       try {
-            System.out.println("Please input name of the movie!");
+        try {
+            System.out.println("Please, input name of the movie!");
             movie = scanner.nextLine();
         }
-       catch (Exception e){
-           e.printStackTrace();
-       }
-       String mov = scanner.nextLine();
-        System.out.println("Please input data about movie!");
+        catch (Exception e){
+            e.printStackTrace();
+        }
+        String mov = scanner.nextLine();
+        System.out.println("Please, input data about movie: ");
         String movie_about = scanner.nextLine();
-        System.out.println("Please input genre");
+        System.out.println("Please, input genre: ");
         String genre = scanner.nextLine();
-        System.out.println("Please input rating");
+        System.out.println("Please, input rating: ");
         int rating = scanner.nextInt();
-        System.out.println("Please input price");
+        System.out.println("Please, input price: ");
         double price = scanner.nextDouble();
-        System.out.println("Please input price for children");
+        System.out.println("Please, input price for children: ");
         double priceForChildren = scanner.nextDouble();
         System.out.println(adminController.createMovie(new Movie(mov,movie_about,price,priceForChildren, rating, genre)));
         forAdmin();
     }
     public void updateMovie(){
-        System.out.println("Please input movie's id which you want to update");
+        System.out.println("Please, input movie's id which you want to update: ");
         int id = scanner.nextInt();
-        System.out.println(adminController.UpdateMovie(id));
+        if (clientController.idExist(id)){
+        System.out.println(adminController.UpdateMovie(id));}
+        else{
+            System.out.println("There is no movie in such id!");
+        }
         forAdmin();
     }
     public void deleteMovie(){
-        System.out.println("Please input movie's id which you want to delete");
+        System.out.println("Please, input movie's id which you want to delete: ");
         int id = scanner.nextInt();
+        if(clientController.idExist(id)){
         System.out.println(adminController.DeleteMovie(id));
+        }
+        else{
+            System.out.println("There is no movie in such id!");
+        }
         forAdmin();
     }
     public void AllMovie(){
@@ -173,27 +185,33 @@ public class MyApplication {
     }
     public void AllUsers(){
         String response = controller.getAllUsers();
-        System.out.println(Objects.requireNonNullElse(response, "There is no accounts "));
+        System.out.println(Objects.requireNonNullElse(response, "There is no accounts!"));
         forAdmin();
     }
     public  void logOut(){
         controller.LogOut();
-        System.out.println("Successfully logged out");
+        System.out.println("Successfully logged out!");
         start();
     }
     public void DeleteUser(){
-        System.out.println("Please input users id");
+        System.out.println("Please, input users id: ");
         int id = scanner.nextInt();
+        if(adminController.UserExist(id)){
         adminController.DeleteUsers(id);
-        forAdmin();
+        }
+        else{
+            System.out.println("There is no user in such id!");
+        }
+        start();
     }
     public void forClient(){
         System.out.println("""
+                ***********************************
                 What do you want to do?
-                1. see all movie
-                2. buy ticket
-                3. return ticket
-                4. log out from my account
+                1. See all movie
+                2. Buy a ticket
+                3. Return the ticket
+                4. Log out from my account
                 5. My items
                 6. Delete my account
                 0. Exit
@@ -206,17 +224,15 @@ public class MyApplication {
             case 4 -> logOut();
             case 5 -> MyAccount();
             case 6 -> {
-                adminController.DeleteUsers(UserController.currentUser.getId());
-                if(UserController.currentUser == null){
-                    System.out.println("User successfully deleted");
-                }
+                System.out.println(adminController.DeleteUsers(UserController.currentUser.getId()));
+                start();
             }
             default -> exit(0);
         }
     }
     public  void allMovie(){
         if(adminController.allMovie().isEmpty()){
-            System.out.println("There is no movie, sorry for this");
+            System.out.println("Sorry, there is no movie! ");
         }
         else {
             System.out.println(adminController.allMovie().toString());
@@ -224,58 +240,54 @@ public class MyApplication {
         forClient();
     }
     public void buyTicket(){
-        System.out.println("Please input movie's id");
+        System.out.println("Please, input movie's id: ");
         int id = scanner.nextInt();
-        System.out.println("Please input date (29/02/2024)");
-        String date = scanner.next();
-        int day = Integer.parseInt(date.substring(0, 2));
-        int month = Integer.parseInt((date.substring(3,5)));
-        int year = Integer.parseInt(date.substring(6,9));
-        if(date.length()!=10) {
-            System.out.println("Incorrect date");
-            forClient();
-        }
-        else {
-            if(!dateExist(day,month,year)){
-                System.out.println("Incorrect date");
-                forClient();
+        if(clientController.idExist(id)) {
+            System.out.println("Please input date (29/02/2024): ");
+            String date = scanner.next();
+            int day = Integer.parseInt(date.substring(0, 2));
+            int month = Integer.parseInt((date.substring(3, 5)));
+            int year = Integer.parseInt(date.substring(6, 10));
+            if (dateExist(day, month, year)) {
+                System.out.println("""
+                            Please, input movie's session
+                              10:30
+                              12:40
+                              14:50
+                              17:00
+                              19:10
+                            """);
+                String time = scanner.next();
+                int hour = Integer.parseInt(time.substring(0, 2));
+                int minute = Integer.parseInt(time.substring(3, 5));
+                if (TimeExist(hour, minute)) {
+                    System.out.println(clientController.buyTicket(id));
+                }
+                else{
+                    System.out.println("Incorrect time!");
+                }
+            } else{
+                System.out.println("Incorrect date!");
             }
+
         }
-        System.out.println("""
-Please input movie session
-  10:30
-  12:40
-  14:50
-  17:00
-  19:10
-""");
-        String time = scanner.next();
-        int hour = Integer.parseInt(time.substring(0,2));
-        int minute = Integer.parseInt(time.substring(3,5));
-        if(time.length() != 5) {
-            System.out.println("Incorrect time");
-            forClient();
-        }
-        else {
-            if(!TimeExist(hour,minute)){
-                System.out.println("Incorrect time");
-                forClient();
-            }
-            ;
-        }
-       if(dateExist(day,month,year) && TimeExist(hour,minute)) {
-            System.out.println(clientController.buyTicket(id));
+        else{
+            System.out.println("There is no such movie!");
         }
         forClient();
     }
     public void returnTicket(){
-        System.out.println(clientController.returnTicket());
-        System.out.println("**********************");
+        if(UserController.currentUser.getId() == 0){
+            System.out.println("You don't have ticket! ");
+        }
+        else
+        {
+            System.out.println(clientController.returnTicket());
+        }
         forClient();
     }
     public void MyAccount(){
         clientController.Myaccount();
-        System.out.println("********************");
         forClient();
     }
     public boolean TimeExist(int hour, int minute){
@@ -284,6 +296,9 @@ Please input movie session
 
     public boolean dateExist(int day, int month, int year){
         if(day<=0 || month <= 0 || year <= 0) return false;
+        if(year<2023 || month<2){
+            return false;
+        }
         int feb = 28 ;
         if(year%4==0){
             if(year%100 != 0){
@@ -293,8 +308,8 @@ Please input movie session
             }
         }
         return switch (month) {
-            case 2 -> day < feb;
-            case 1, 3, 5, 7, 8, 10, 12 -> day <= 31;
+            case 2 -> day <= feb;
+            case 3, 5, 7, 8, 10, 12 -> day <= 31;
             case 4, 6, 9, 11 -> day <= 30;
             default -> false;
         };
@@ -312,11 +327,12 @@ Please input movie session
             }
         }
         if(n>0 && n<8){
-            throw new IncorrectPassword("Your password should contain at least 8 digits");
+            throw new IncorrectPassword("Your password should contain at least 8 digits! ");
         }
         if(n != cnt){
-            throw new IncorrectPassword("Your password should contain only numbers and character");
+            throw new IncorrectPassword("Your password should contain only numbers and character!");
         }
     }
 }
+
